# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Ingredient {
  item: string;
  /** @nullable */
  note?: string | null;
}

export interface InstructionStep {
  step: number;
  text: string;
  /** @nullable */
  timerMinutes?: number | null;
}

export interface Nutrition {
  /** @nullable */
  calories?: number | null;
  /** @nullable */
  proteinG?: number | null;
  /** @nullable */
  carbsG?: number | null;
  /** @nullable */
  fatG?: number | null;
}

export interface RecipeSummary {
  id: number;
  slug: string;
  title: string;
  description: string;
  category: string;
  imageUrl: string;
  /** @nullable */
  prepMinutes: number | null;
  /** @nullable */
  cookMinutes: number | null;
  /** @nullable */
  totalMinutes: number | null;
  dietaryTags: string[];
  cookMethod: string;
  rating: number;
  reviewCount: number;
}

export interface Recipe {
  id: number;
  slug: string;
  title: string;
  description: string;
  category: string;
  cuisine: string;
  imageUrl: string;
  /** @nullable */
  prepMinutes: number | null;
  /** @nullable */
  cookMinutes: number | null;
  /** @nullable */
  totalMinutes: number | null;
  /** @nullable */
  servings: number | null;
  /** @nullable */
  sourceUrl?: string | null;
  dietaryTags: string[];
  cookMethod: string;
  rating: number;
  reviewCount: number;
  ingredients: Ingredient[];
  instructions: InstructionStep[];
  nutrition: Nutrition;
  createdAt: string;
}

export interface Category {
  name: string;
  imageUrl: string;
  recipeCount: number;
}

export interface SubstitutionRequest {
  /** @minLength 1 */
  ingredient: string;
  note?: string;
}

export interface SubstitutionResponse {
  ingredient: string;
  suggestion: string;
}

export interface RatingRequest {
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
}

export interface RatingResponse {
  rating: number;
  reviewCount: number;
}

export type ListRecipesParams = {
/**
 * Free-text or natural-language search, e.g. "chicken thighs and 20 minutes"
 */
query?: string;
category?: string;
sort?: ListRecipesSort;
};

export type ListRecipesSort = typeof ListRecipesSort[keyof typeof ListRecipesSort];


export const ListRecipesSort = {
  popular: 'popular',
  recent: 'recent',
} as const;

export type ListPopularRecipesParams = {
limit?: number;
};

export type ListRecentRecipesParams = {
limit?: number;
};